Charm: Remove core/charm format
The following removes core/charm format in favour of charm/v8 format.
This is so we can keep the format logic in the same place as the charm.
Ideally we should never have to care about the format in juju code, but
unfortunately we don't have the option and would require a lot of
re-writing for maybe not much reward.