Upgrades: Ensure we have an empty manifest from state
In order to have better determination of when a charm meta data format
is either v1 or v2, it helps if we don't have charms that have empty
manifests. Empty manifests with empty bases means the logic for
determine if it's a format v2 now has to check for the presence of
values rather than checking for nil.

Lots of code to just say `version: 2` in a metadata.