Logging: Reset optional fields
When the logsink reads the log records on the connection, it reuses one
log record for all reads. The problem occurs when there is a partial log
record which can have missing entities or labels. As entity and labels
are optional; if they happen to be set, but the next one doesn't have
them, then the ReadJSON (unmarshaller) doesn't clear ALL the fields, it
just updates what is required. This results in labels from the previous
read being in all future reads, until a new record happens to have a new
entity or label.

The code is simple, reset all the optional fields so that we always
start with a fresh log record. We don't want to create a new LogRecord
for every read as that's a performance regression.