Add helper for looking up the origin of an uploaded charm via its URL
When attempting to store a charm into the blobstore, we first check if
the data has already been uploaded. If so, the store is a no-op but we
need to return back the recorded origin stored by the application that
references the charm URL that we were asked to store.

This helper allows us to perform a reverse lookup (from URL to the
application) and extract the origin. To reduce further coupling to the
state package, the commit adds a conversion helper to the state.Origin
type which transforms it into a core/charm.Origin value.