Add CHARMHUB logger label.
Use it when creating a charmhub client to faciliate debugging.

Ensure CHARMHUB logging label is used. Swap the logger pattern used in the
charmhub package. Create a child logger and give a label inside of charmhub,
to ensure that all loggers passed in get the benefit of the label.  Callers
of CharmHubConfig and CharmHubConfigFromURL should just pass their logger.