Refactor charm lookup by URL method in state
The original implementation of the lookup code returns ErrNotFound for
placeholder and/or charms pending to be uploaded. However, the
asynchronous charm downloader code will not use the two-phase commit
approach of the current code (i.e. prepare the charm doc entry, download
charm, update the entry with both the metadata *and* the stored blob
path/SHA256) but rather inserts a charm document with the metadata (but
no blob details) and the pending-download flag set.

To this end, the lookup code has been modified to inspect the controller
flag list and allow pending charms to be retrieved if the async charm
download flag is set. This will allow the rest of the existing deploy
steps to function properly.