Add escape hatch to AddCharmXXX to trigger an async charm download
The escape hatch is behind the async charm downloads feature flag. When
enabled, the AddCharm and friends will fetch the relevant charm metadata
and persist it as a new charm document with the pending flag set.

Currently, the repository implementation do not provide the means to do
a batch give-me-the-charm-meta-only request so the call simply emulates
this behavior by downloading the charm (as the current AddCharm
implementation does) and reseting the calculated SHA256 value (note: the
code uses the repository DownloadCharm method directly (instead of the
DownloadAndStore abstraction) so no blobs are stored.

The server-side-bundle expansion work will (eventually) provide its own
dedicated facade (the batch metadata download functionality will be
implemented then) thus rendering the methods in the charms facade
obsolete. However, the use of the controller flag allows us to test the
async charm download functionality *now* without breaking the existing
deploy flow.