Pass through notify target to global clock
The following passes through the notify target for the global clock
updater. This is an interim state, as we actually want to remove the
raftforwarder completely or how the communication is setup between the
leasemanager and the raftforwarder.

Part of the path to doing that is lifting the lease log construction
out of the raftforwarder manifold and into the machine agent config
construction. Then we can pass it to both the globalclockupdater and
the raftforwarder workers.

There are still tests for the globalclockupdater updater code, which
will follow.

----

The communication pattern between the leasemanager and the raftforwarder
in the current code is weak. Pubsub doesn't feel like the appropriate
callback mechanism to distribute strong consistency communication
between two workers in the same controller. Although raft should deal
with faults from missing events, it feels like we shouldn't be
attempting to introduce faults wilfully.