Lease: Abstract out the pubsub
The following changes move the pubsub requests out of runOnLeader into
a client. That way we can swap out the implementation for the lease store
without knowing about the client implementation.

This should make the lease store much simpler because you don't have to
workout how the pubsub publish and subscribe setup works.