Pubsub: Add metrics for centralhub
The centralhub is by the very definition central to Juju. So the
following adds metrics to see what's happening at runtime. We can use
these metrics to better understand if pubsub is being jammed up at any
point and if it is, we can then see if it also coincides with raft also
blowing up.

The implementation is really simple, just pass a new metrics type into
the central hub and provide ways to collect the metrics at the right
time.