Worker: Lease consumer
The following is the scaffolding for a new worker. The lease consumer is
the inverted raft forwarder. The lease consumer will only run on a
leader and will accept connections that will be processed. The command
operations payload will then be run on the read-write raft FSM entity
target.

That way we can drop all the pubsub calls between controllers. This
should simplify A LOT of communications between controllers and also
improve reliability.