Ensure unsubscribe on error
If we subscribe to a given topic, we need to ensure that we unsubscribe
when we get an error. If we don't do this, then we can end up having a
instance hanging around for ever. This could end up being a leak of
resources.

Unfortunately pubsub is never tied to lifecycle of an instance, it would
be nice to either pass in a tomb or a context to do auto unsubscribing
once a lifecycle is complete.