Make the revno threshold for the collection watcher configurable
Prior to this change, the collect() function which the collection
watcher invokes would flag the ID of documents that have a revno <= 0 as
not present. As a result the mergeIDs function which is subsequently
called to figure out which changes should be emitted by the watcher
would ignore those documents.

When a new document gets inserted into a collection it gets assigned a
revno of 0. As a result, the collection watcher will not report it even
if the configured filter function returns true for the document. By
making the revno threshold configurable, we can work around this
limitation.

The default zero-value of the threshold allows us to retain the original
watcher behavior and only override as needed to support special
use-cases.