Support marshaling/unmarshaling of NotYetAvailable apiserver errors
The API server will map NotYetAvailable errors to a 409 (conflict)
status code. As per the spec (see
https://www.w3.org/Protocols/rfc2616/rfc2616-sec10.html#sec10.4.10):

"The request could not be completed due to a conflict with the current
state of the resource. This code is only allowed in situations where it
is expected that the user might be able to resolve the conflict and
resubmit the request".