RaftLease: Introduce a new controller facade
The new controller will allow the changing of a lease and apply it to
the notification target if it's a leader. If the controller isn't a
leader it will give as much help about what is the leader and return
that information.

In order to make this work, we have to pass a lot of things into the
apiserver directly and ensure that the manifold correctly starts up.

----

The design of the interface for the raft access has been restricted on
purpose, with the intention that no facade can get access to the
underlying raft instance. Instead we offer a raft mediator. The mediator
negotiates between the various states of raft before attempting to apply
the command.

Because of this we only open a very small change to the apiserver from
the facade side. It would be nice to be able to validate a command
before attempting to apply it, that way we can ensure it's a command and
not some random garbage.

----

One further thought, we also now get the strong hardening of auth,
rather than the potential weak loop hole that is employed via the /raft
endpoint.