RaftLease Client
The raft lease client uses an api client to talk to a facade. The client
holds a series a remotes that is assigned an address. If the address
changes then we update the client and give it a new connection. All
messages will be sent to a remote and if the remote isn't the leader
then attempt to find the new.

The complications arise because we can't get the raft instance into the
APIServer as that causes dependency.ErrMissing. We've unfortunately
created as circular dependency issue here and I'm unsure how to unwedge
this.

Note: the client requires more work as we can fully test this end to end
yet.