Network devices from config host_name
The following changes ensures that the eth0 device gets a unique name.
This is given to us via LXD. Unfortunately we're unsure currently why
you would have interfaces first time around and then devices another
time around, which seems to be the only way to trip this up?

More information from the host machine would be required to find the
underlying bug `ip -a` and a list of all the nics.

For now this seems like it has the potential to unblock containers being
created.