Remove UploadedCharmOrigin method from state
The method was introduced in https://github.com/juju/juju/pull/13191
but was rendered useless after the deploy specific revision changes
landed in https://github.com/juju/juju/pull/13206.

Since (on charmhub) the multiple channels may use the same revision to
refer to the same charm blob, we should never pull the recorded origin
value from an existing application but always resolve the origin on
demand.

For example:
$ juju deploy redis # deploys revision 42 from stable
$ juju deploy redis --revision 42 --channel 2.0/stable another-redis

In the above example, both "redis" and "another-redis" reference the
same charm URL but on different channels so resolving the origin is now
required so that we always report the appropriate origin.