Skip badly written e2e tests in BundleDeployCharmStoreSuite
These tests are full e2e tests that invoke external APIs (charmstore)
which we cannot really mock as the client has no means to control the
initialization of the client/charms facade (the AddCharm endpoint in
particular).

The issue was uncovered due to the fix introduced by the previous
commit.

Simply mocking the API calls defeats the purpose as the actual test
expectations typically query state documents which in turn means that
the tests must be e2e and therefore mocks would not be applicable.

To this end, the entire test suite has been skipped.