RemoteConnectionStatus: KeyRelation not found
When attempting to calculate the RemoteConnectionStatus in the
AllWatcher, it is not the time to bomb out when you can't find a
KeyRelation. As the RemoteConnectionStatus only wants to know the active
count, the very fact that you can't find the KeyRelation indicates that
it's no longer active. Ignoring the error and continuing onward should
suffice.

Other code found at[1] also deals with the error in exactly the same
way.

 1. apiserver/facades/client/client/status.go#fetchOffers