Update logic in applicationsForMachineChange.
applicationsForMachineChange relies on the old sort algorithm for
changes.  The new topo sort algorithm breaks it, causing an infinite
loop due to its recursive nature.  Instead use changes rather prior
change handling to determine what units will be added to a machine or
container.

When calling resolve, check that the item we're looking for exists,
has been handled already.  If not, there is an error in the requirements
for a change.