Use CIDR in CIDR fallback when matching machine addresses to subnet CIDRs
Some providers (e.g. equinix) carve out a portion of an available subnet
and assign addresses off that to provisioned machines. For example, the
subnet pool might be a /25 and machines might get assigned a /31 carved
off the /25 range.

As a result, when executing network-get, Juju is unable to match the
subnet CIDRs reported by the machine agent to a subnet CIDR (the test
uses CIDR equality as its predicate) and therefore omits otherwise
usable private addresses from the reported ingress address list.

To fix this, SubnetInfos.GetByCIDR has been updated with a fallback: if
no match has been found using the equality check, a second match attempt
will be performed where we check whether the provided CIDR value (from
the machine address) is a subset of the subnet CIDRs that we know of.

To perform the CIDR-in-CIDR check, we first calculate the first and last
IP (inclusive) in the CIDR value passed to GetByCIDR. Then, for each
known subnet, we check whether its net.IPNet representation containts
*both* IPs.