Fake GetEssentialMetadata for charmhub repository
The proposed method implementation groups together requests based on the
macaroons provided (requests without macaroons are lumped together) and
attempts to fetch the data in batches (note that macaroons are not
currently supported by charmhub but the repository is wired to use them
once support becomes available).

At the moment, since the charmhub API (via the refresh method) only
allows us to access the metadata.yaml contents, the implementation falls
back to downloading the charms and populating the essential metadata
responses.

When the charmhub API rolls out support for fetching the metadata we
need (metadata.yaml, manifest.yaml, config.yaml, lxd-profile.yaml), we
will amend the implementation accordingly.

For the time being nothing is calling this particular repository method
so it is safe to land so we can try out the asynchronous charm download
flow (behind a feature flag).