Adds support for attaching Instance Profiles.
This change introduces the ability to attach already made and new
Instance Profiles to Juju controller machines using Bootstrap
constraints.

With this change a user can specify and already existing instance
profile with --bootstrap-constraints="instance-role=test" or
--bootstrap-constraints="instance-role=auto" for the Juju bootstrap
process to make a brand new instance profile in the form of
juju-controller-<controller-name>.

This PR does not make the controller use these new credentials it just
introduces the attachment and creation process.