Update async charm d/l code in charms facade to only fetch essential metadata
The initial (placeholder) implementation of the queueAsyncCharmDownload
method in the charms facade downloaded the charm and extracted the
required metadata from it.

With this commit, the implementation now uses the GetEssentialMetadata
method from the charm repository to grab the minimum set of metadata
needed (metadata.yaml, manifest.yaml, config.yaml and lxd-profile.yaml)
for a deployment without having to download the full charm and pass
those to state.AddCharmMetadata.

The remaining bits of metadata (e.g. actions, metrics etc.) are
populated once the charm gets asynchronously downloaded and before the
charm blob becomes available for agents to download.