Github action for windows
The following adds github windows action for client only tests. As we're
only testing clients only and not workloads, we don't _really_ need the
correct version of mongo here. All full end to end tests should be
officially identified and either moved to a better location or
re-written to prevent full stack testing. The CLI shouldn't be
exercising the full stack, that if anything should be apiserver and
down.