lxd: Update IsLXDNotFound to deal with Go HTTP spelling
LXD is now using Go's native HTTP error which come with different
capitalization, so update to support both.

In the future, it'd be preferable for Juju to gain support for LXD's
StatusError type which allows for direct comparison with HTTP errors as
well as retrieving the HTTP error code directly.

In general string matching on errors is a very bad idea and the LXD team
would like to get notified when that's necessary so we can come up with
an alternative.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>