Add ErrDropped to indicate when messages are dropped
Messages can be dropped when there is no way to forward them on to a
leader. This allows us to split the error messages from ErrTimeout.
Previous implementations just dropped the messages on the floor and
returned no error, so it was assumed that a claim happened, even though
the claim never actually stuck.

Sad times.