Remove juju generated host_name
The host_name should be generated directly from LXD, instead of Juju.
We only set the host_name to identify the devices/nics when looking at
them via `ip a`. As this was just a convenience, we should fallback to
letting LXD do it. This will prevent naming collisions at the kernel
level, where they haven't gc'd the devices in time.