Raft worker errors
In order to improve readability of why a raft worker is restarting, the
following introduces logging error messages.

The problem isn't helped by the masked ErrStartTimeout waiting for a new
raft channel. As the catacomb is killed by the ErrStartTimeout, we loose
the original error message.

We should revisit if this is the best way to do this!