Lease Manager: Dropped Operation
The following exposes a dropped operation error inside the lease
manager. Both pubsub client and the new up and coming raft lease client
have the potential to drop requests.

The dropped requests are a manifestation of when a connection doesn't
exist to a remote server (controller). Attempting to connect to the
remote server in a ad hoc fashion (lazy), is problematic. The
operational lease generally have small timeouts (5 seconds) and
connection to a given remote server can extend up to minutes (worst
case). Instead we should allow the clients to setup a connection in the
background and if there isn't one return an error that the lease manager
can work with. The pubsub client will still silently drop operations and
return nil, a timeout waiting for the result then be trigger as there
isn't a way to apply back pressure easily. Future PRs will address the
the ErrDropped by applying more delay, to ensure that the lease manager
isn't requesting leases on a leader with no connection.

The code for this PR is simple, introduce a new sentinel error and
handle the result of the error in the lease manager. As nothing is
triggering this error atm, it shouldn't impact any running operations.