Raft queue: Blocking queue
The following adds a blocking queue for apply raft commands. The queue
is a blocking queue to force back pressure on to the enqueuing side.
This is the opposite of the current implementation of applying raft
leases, where the pubsub client can't doesn't apply back pressure for
all applications, as the very nature is fire and forget. Workarounds for
this including subscribing to a one of event topic to see the result,
although that takes a lot of time to wait for the result.

The code is simple, wrap two channels, that allow the enqueuing of
operations and wait for the error channel for the result. It's then up
to the raft worker (to be following in another commit) to consume the
queue and to enqueue the error accordingly.