Raft worker: Apply raft operation
The following adds a function for applying raft operations to the raft
log. If the raft instance isn't the leader, then a NotLeaderError is
returned to the operational queue. The leader information might not be
available, in that instance it is appropriate to send empty information
and it's up to the client of the queue on how to handle that.

The Raft spec[1] outlines this "The first option, which we recommend
..., is for the server to reject the request and return to the client
the address of the leader, if known." (see 6.2.1 of the spec[1]).

In future commits/PRs, the APIServer will append items to the queue so
that we can correctly follow the spec.

The code is rather mechanical, attempt to determine the state of the
raft instance, if it's the leader apply the command to the FSM,
otherwise return an error.

1. https://web.stanford.edu/~ouster/cgi-bin/papers/OngaroPhD.pdf