Prevent deadlock
If the context timed out on the reading of the error, we might deadlock
in the enqueue. Which we would never want to happen. In this case, we
always want to read the error, every time.

Additionally, this removes the context and uses a deadline to for the
operation. That way we can use the enqueuing time as part of the timeout
for an operation. This should reduce the worst case scenario of being 10
seconds for enqueuing, not including network transports.