Worker: Wire up apply and queue
Wires up the worker to consume items of the queue and apply them to the
raft instance. As only one item can be processed of the queue at any one
time, then we serialize the applications to raft. We can then apply back
pressure to anyone attempting to call a lease.

The dependency engine isn't wired up yet, that comes later.