APIServer: Raft context
The raft context is a way to mediate between the RaftOpQueue with a
worker listening on the other end and a way for facades to enqueue
commands onto the queue. It's done like this, so it doesn't expose the
raft instance arbitrary to the facades, limiting abuse and side effects.

Additionally, having a queue in between the instance and the facade
allows us to apply back pressure and deadlines to enqueuing.

The code is simple, but _very_ mechanical, as we have to thread the
queue from the machine manifolds all the way to the facade context.