RaftLease Client: Handle lease requests
The following adds a raftlease client. The aim is to provide a robust
way of sending requests whilst being honest about the error messages
that fail. In particular we drop requested commands when we don't have a
connection and it's up to the callee to retry again. Previous
implementations just dropped them on the floor.