Facades: Handle not leader errors
The following handles the not leader error differently from a normal
error. If we get a not leader error, we should stop processing any more
leases and mark all subsequent errors as the same and then exit out.
Normal errors, we do want to keep iterating on, as we just don't if we
can bail out early unless we have more typed error cases.