Update lxdprofile watcher to emit changes when charm metadata changes
When the asynchronous charm download flag is enabled, the charms client
facade will only populate the essential bits of the charm document which
are required to commence deployment (metadata, manifest and config).

The full charm blob is then downloaded asynchronously while the machine
for the application units are being provisioned and the full metadata is
populated once the download completes.

Any potential lxd profile defined by the charm will only become known to
Juju once the download is complete. We must therefore rely on the
instancemutator worker to apply the lxd profile to the relevant
containers.

In order for the instancemutator to pick up this change, it needs to
watch for charm metadata changes, match the charm URLs to the ones
referenced by the applications whose units are assigned to a particular
machine and then compare the LXD profiles for changes. This commit
introduces this extra bit of functionality.