Temporarily add a replace directive for gocheck to log failures
For some reason, occassionally we get a test failure that just says
"1 FAILED" at the end of the suite, but no indication or logs of which
test failed.

This is shown in the following CI run (with just the relevant suite
extracted into a Pastebin):
https://jenkins.juju.canonical.com/job/github-make-check-juju/12888/consoleFull
https://pastebin.canonical.com/p/yfsRNNnzpG/

It's possible something is going to the logs and we're not seeing it
because -check.vv is not enabled, but this will allow us to detect it
and narrow it down to a specific test (as well as see any logs to that
point).

The gocheck diff is shown here:
https://github.com/benhoyt/check/commit/a82805bc27e8389b59cc95b6061e3302181f5753