Pass HTTP Client through
The following changes passes in a http client that can be used to
improve retries in certain situations. The status codes that are
retryable are defined in juju/http/v2;

 - http.StatusBadGateway
 - http.StatusGatewayTimeout
 - http.StatusServiceUnavailable
 - http.StatusTooManyRequests

All other status codes are deemed unretryable from a server and HTTP RFC
perspective.

The code is simple, inject a http.Client in the provider when generating
a LXD remote server and LXD will automatically pick this up.