State: Logs already exist
When creating a model with a given UUID it is expected that if the logs
already exist then don't throw an error. Instead check the error and
return nil. Although it would be great to use a typed error here, we
aren't fortunate to get it from the mgo library. For now, we'll test for
the error code directly, we can be sure it's right by spelunking the
mongo source code.