InstanceKey for Origins.
Related to charmhub KPIs and LP: 1944582. Provide a stable, unique InstanceKey
when using the refresh action for charmhub. The easiest one is to combine the
model UUID and the application name.

Setup for most calls in GetCharmURLOrigin, to be used when creating a
RefreshConfig down in refreshOne.  All client initiated refresh actions end up
there.  However the data required isn't easily available.

The charmrevisionupdater refresh action is put together in a facade and
has easy access to the model UUID and application name.

For compatibilty with older clients, if no InstanceKey is provided to
RefreshOne, create a uuid.

Do not add to state's origin.