No need to error here for not found networks
So we get the network names from the lxd server, and then start
querying the server with those names one by one for the network
details. Sometimes right between these two operations a network goes
away and the server reports that it hasn't got that network anymore
(#raceisreal). This happens very often in pylibjuju's integration
tests on CI. Here it makes sense to log a warning instead of erroring,
since this is not actually an error.