Batching FSM
The following introduces the BatchingFSM, the idea behind this, is to
reduce lock contention and allow for applying log changes in chunks.
This is intended to work out if there is any performance gains by using
this with either transport (API or pubsub).

I suspect when load testing that we won't actually see any changes for
the API version, as we intentionally serialise there. So changes to
consolidate multiple logs at once will be required. The pubsub one
should actually go faster as we don't have to wait for the log to be
applied.