Add updated JIMM facades
It looks like the schema gen was missing the latest controller facade
versions for schema gen. It also looks like we need to drop older facade
versions for JIMM and this will require some coordination.

For now, we'll bump the facade versions and deal with the coordination
for a later date.