Lease: Coalesce lease applications
The following is very experimental and the idea is to coalesce the lease
operations into batches that raft can work on. They either get enqueued
as a series of logs via the amount or they are enqueued after 100
milliseconds.

For the moment the coalesce amount is 1, but during performance testing
a higher amount might be very useful.