Add upgrade step to drop assumes keys in charm metadata
Unfortunately, we rushed to add the "assumes" section (a []string) to
the charm metadata before the spec was finalized and even when no charms
out there supported it. The assumes spec specifies an expression tree
syntax which is incompatible with what older controllers would be trying
to shove into the database.

As a result, when charms with assumes block do become available, older
controllers will insert garbage to the database (or simply refuse to
parse the charm metadata as the new syntax can not always be coerced
into a []string).

This, plus the fact that we directly serialize the charm metadata (from
the charm package) into the charms collection instead of first mapping
it to a struct within the state package now means that we need an
upgrade step to drop existing "assumes" keys from the charms collection
so we can populate the field correctly with the parsed expression tree
from 2.9.17 and onwards.