Relax validation constraints when changing the default arch for a model
The original code in state.constraintsValidator() used to enumerate the
known set of arches by inspecting the (downloaded) cloud metadata and
any manual machines associated with the model and use this information
to override the arch-specific vocabulary used by the returned
constraints validator.

This seems to strict as the underlying substrate may still be able to
provide instances with an arch we haven't used so far.