Alternative coalescing queue
The following implements a coalescing queue using two channels, which
allows for better coalescing semantics.

There is a naive backoff algorithm to allow for better optimistic
batching. Although just thinking about it now, we might actually want to
reduce the number of operations that can be added the further out the
delay, so that raft can recover.