Ensure the client handles leader retries only
The following ensures that we only retry at the client level if the
leader changes. All other errors will be bubbled up to the lease
manager. This we we prevent the double retries happening at every level.

We also make sure that we don't drop the lastKnownRemote in every
situation.