Expose raft apply metrics
These metrics report exactly the time it takes to apply an operation to
the raft log. Instead of using the client timings, which also factor in
network latency, the following ensures that we just call out the raft
log times. This is important when trying to understand where the cause
of slowness is.