Do not close stdout when restarting the apiserver
The API server when killed will attempt to clean up any dependencies it
has created in it's lifetime. Except that one of those for the CAAS
model type is stdout. If you close stdout you can no longer write to it.

Turns out, our tests exercised this code path and we closed stdout
preventing the writing of any test descriptions to stdout. The tool to
analyse the output of the log then assumed there was a panic and marked
all our tests in this package as failed. Why it wasn't reported as a
panic in Jenkins was because the test coverage tool only looked for the
PANIC key phrase.

This was PAINFUL to debug...