Prevent the raft queue from panic'ing
The raft queue under extreme load will timeout when attempting to shed
load. When this happens, an error is returned informing the caller that
a timeout occurred. If the closure returns after the fact, then
attempting to send on a closed channel causes the worker to panic.

Additional work includes removing the need to write to a buffered
channel, instead we just queue all the enqueues up until it's possible
to dequeue them. This allows for a better connection ratio and we can
still shed the load by bouncing them if the timeout occurs.