Fix flaky worker/caasrbacmapper tests (and remove time.Sleep calls)
The test flakiness was in TestRBACMapperUpdateSync, and was caused by
the test sometimes calling mapper.AppNameForServiceAccount before the
worker had a chance to add it. A couple of the tests used time.Sleep to
work around this, but one test already had a more reliable "attempt
loop" for this, so I've copied that pattern to all tests.