Remove the raft lease log
The raft lease log is currently used to gather information during a
notify to the notify target. Essentially the aim is to report when
a lease is claimed or revoked, but also if a failure to perform a
transaction in mongo.

There are problems with this setup:

 1. Data is duplicated in 3 locations:
     - machine log
     - mongo logs
     - raft lease logs
 2. The raft lease logs require compressing and log rotation if they're
 over a given size.
 3. The performance of claiming is slowed down because of the additional
 writing to the logs.
 4. Does anyone actually use or know about them

All this means that we've added a lot of complexity and duplication when
it's not generally required. Discussing with others, it's better to just
remove the lease log and ensure that we do log out to the other
locations.

Finally, if mongo is struggling for writes, the machine log will have
the information. So replication is performed at the small locality,
enough to be able to debug.