Remove controller API from deploy
The following removes the controller API connection call from deploy,
which is required for JAAS to work when deploying charms. Only model
level API connections are allowed not controller API root, as users
aren't allowed to access those from a permission perspective.

When making the change it makes sense to go and remove all references
and fallback to the clients handling of the URL. The operator doesn't
have the  ability to set the charmstore URL. The charmstore URL isn't
going to change and it's not possible to stand your own charmstore up
anyway. If no URL is passed into the client, then the charmstore client
will use the csclient.ServerURL, which is what we would do anyway.

We can leave it in the controller config for now, as we can write a
migration step to remove that at a later date.

This does simplify the construction of the store as we no longer pass
around the charmstore URL, but all the tests need updating.