Batch expire the expired leases
This is forerunner to batching all the notify targets. Essentially this
a best effort of removing a lease after a revoke. The next step would be
to batch all claims as well, but that can be done in another PR.

The following should easily benefit both pubsub and API raft lease
transports.