Add num-provision-workers configuration option
This option specifies the number of workers used by the provisioner code
to process start/stop requests. This value enforces an upper limit on
the number of machines that can be spinned up concurrently but no limit
on the number of machines that can be terminated as those happen in
batches.